\name{coef.CARBayes}
\alias{coef.CARBayes}


\title{
    Extract the regression coefficients from a model.
}
\description{
    This function takes a CARBayes object and returns the vector of estimated
    regression coefficients (posterior medians).
}

\usage{
    \method{coef}{CARBayes}(object, \dots)
}



\arguments{
    \item{object}{
        A CARBayes fitted model object.
    }
    \item{\dots}{
        Ignored.
    }
}
\author{
    Duncan Lee
}
