\name{dbf}
\alias{dbf}

\docType{data}

\title{
Shapefile (.dbf) for all Intermediate Geographies (IG) in Scotland.
}

\description{
dbf contains the .dbf component of the shapefile for all Intermediate Geographies (IG) in Scotland, which is used in the worked example in the main help page of the CARBayes package.
}

\usage{data(dbf)}

\source{
The shapefile was provided by the Scottish Government, and can also be downloaded from http://www.sns.gov.uk/. Further details about this shapefile can be found at the same location.
}

\examples{
data(dbf)
}

\keyword{datasets}
