\name{fitted.CARBayes}
\alias{fitted.CARBayes}


\title{
    Extract the fitted values from a model.
}
\description{
    This function takes a CARBayes object and returns the vector of fitted
    values (posterior means).
}

\usage{
    \method{fitted}{CARBayes}(object, \dots)
}



\arguments{
    \item{object}{
        A CARBayes fitted model object.
    }
    \item{\dots}{
        Ignored.
    }
}
\author{
    Duncan Lee
}
