\name{logLik.CARBayes}
\alias{logLik.CARBayes}


\title{
    Extract the estimated loglikelihood from a fitted model.
}
\description{
    This function takes a CARBayes object and returns the estimated loglikelihood
     (posterior means).
}

\usage{
    \method{logLik}{CARBayes}(object, \dots)
}



\arguments{
    \item{object}{
        A CARBayes fitted model object.
    }
    \item{\dots}{
        Ignored.
    }
}
\author{
    Duncan Lee
}
