\name{model.matrix.CARBayes}
\alias{model.matrix.CARBayes}


\title{
    Extract the model (design) matrix from a model.
}
\description{
    This function takes a CARBayes object and returns the design matrix.
}

\usage{
    \method{model.matrix}{CARBayes}(object, \dots)
}



\arguments{
    \item{object}{
        A CARBayes fitted model object.
    }
    \item{\dots}{
        Ignored.
    }
}
\author{
    Duncan Lee
}
