\name{print.carbayes}
\alias{print.carbayes}


\title{
Print a summary of the fitted model to the screen.
}
\description{
This function takes a carbayes object and returns a summary of the fitted model.
}

\usage{
\method{print}{carbayes}(x, \dots)
}



\arguments{
  \item{x}{
A carbayes fitted model object
}
  \item{\dots}{
Ignored
}
}
\author{
Duncan Lee
}

\examples{
## See the examples in the main help file for CARBayes.
}