\name{print.carbayes}
\alias{print.carbayes}


\title{
Print a summary of a fitted carbayes model to the screen.
}
\description{
This function takes a fitted model object of type `carbayes'  and returns a 
summary of it.
}

\usage{
\method{print}{carbayes}(x, \dots)
}



\arguments{
  \item{x}{
A `carbayes' fitted model object
}
  \item{\dots}{
Ignored
}
}
\author{
Duncan Lee
}

\examples{
## See the vignette accompanying this package for an example of its use.
}